requirements.txt
hydra-core
lightning
torch
numpy
transformers
datasets
tiktoken
wandb
tqdm
yaml
